new changes

+++ setup_project.sh
@@ -12,10 +12,14 @@ mkdir -p "$project/Helpers"
 mkdir -p "$project/reports"
 
 # Moving existing source files into correct structure
-mv attendance_checker.py "$project/"
-mv assets.csv "$project/Helpers/"
-mv config.json "$project/Helpers/"
-mv reports.log "$project/reports/"
+[ -f "attendance_checker.py" ] && mv attendance_checker.py "$project/"
+[ -f "assets.csv" ] && mv assets.csv "$project/Helpers/"
+[ -f "config.json" ] && mv config.json "$project/Helpers/"
+[ -f "reports.log" ] && mv reports.log "$project/reports/"
+
+#Archiving and cleaning Interrupted project
+trap 'echo "Interrupted! Archiving project..."; tar -czf "${project}_archive.tar.gz" "${project}";
+rm -rf "$project"; exit' SIGINT
 
 # Ask if user wants to update thresholds
 read -p "Do you want to update thresholds? (yes/no): " choice
@@ -23,12 +27,12 @@ read -p "Do you want to update thresholds? (yes/no): " choice
 if [ "$choice" = "yes" ]; then
 
     # Get new values from user
-    read -p "Enter warning threshold: " warning
-    read -p "Enter failure threshold: " failure
+    read -p "Enter new warning threshold: " warning
+    read -p "Enter new failure threshold: " failure
 
     # Update config.json safely
-    sed -i "s/\"warning\": [0-9]*/\"warning\": $warning/" "$project/Helpers/config.json"
-    sed -i "s/\"failure\": [0-9]*/\"failure\": $failure/" "$project/Helpers/config.json"
+    sed -i "s/\"warning\": [0-9]*/\"warning\": $warning/" "$project/Helpers/config.json/"
+    sed -i "s/\"failure\": [0-9]*/\"failure\": $failure/" "$project/Helpers/config.json/"
 
 else
     echo "Keeping default threshold values"
@@ -41,13 +45,5 @@ if python3 --version > /dev/null 2>&1; then
 else
     echo "Warning: Python3 not found "
 fi
-
-# Handle Ctrl+C interruption
-
-trap 'echo "Interrupted! Archiving project...";
-tar -czf "${project}_archive.tar.gz" "$project";
-rm -rf "$project";
-exit' SIGINT
-
